agents/requirements.txt
anthropic>=0.39.0
python-dotenv>=1.0.0
schedule>=1.2.0
pydantic>=2.0.0
rich>=13.0.0
